[RealSenseRecorder.py]
- Adding a new function in order to analyze and calculate the depth 
coverage of the depth frames dataset (getDatasetDepthQuality).

- Adding post-processing effects, allowing to increase the quality of 
the recorded depth frames.

[Reconstructor.py]

- Fixing a bug which was preventing some frames to be counted and used 
during the generation of the last shard.

[ShardProcessor.py]

- Enhancement of the process logs status.

- Adding the "modulo 5 case" registration mode in order to strengthen 
the registration accuracy in case of loop closures.

[ShardAssembler.py]

- Refining the registration of the different shards using a 
point-to-plane registration following the RANSAC global registration.

- Adding global optimization on the posegraph before merging the shards.

[Logger.py]

- Fixing a mistake which was causing the multiple and excessive 
initialization of the colorama library.

[VirtualVision.py]

- Adding the dataset depth quality analysis following the scan.

[README.md]

- Adding the OpenStack cloud deployment as a new intermediate goal.

- Removing the Unity3D importation goal as Open3D currently provides a 
3D viewer which seems to fit the needs.

[Progress]

- Some work has been done in order to add the compatibility with the 
Kinect V2. However, the development of such a feature will not be 
continued because of the deprecation of multiple Kinect libraries and 
the difficulty to align the RGB and depth frames. It might be considered 
once again in the future.

- A lot of work has been done in order to do the camera calibration in 
order to virtually set the RGBD camera (for the detection) location 
within the 3D model. A large commit is expected soon.