[/Client/Virtual_Vision.py]
- Adding a camera calibration mode in order to semi-automatically set 
the camera's real location within the "World" point cloud

- Adding a real-time object detection and 3D integration mode, which 
allows to place the objects detected by the RealSense camera into the 
world point cloud at their exact real location

[/Client/config.json]

- Adding a JSON configuration file in order to set the parameters for 
the calibration and detection mode. This file will be used in future as 
an alternative to CLI arguments.

[/Client/Calibration/Calibrator.py]

- This class has been developped in order to semi-automatically 
determine the camera's position within the "World" point cloud using a 
"POV" point cloud made from a set of RGBD image from the camera point of 
view.
It is possible then to calculate the transform in order to move the 
objects detected into their correct location in the "World" point cloud

[/Client/Detection/darknet.py]

- Adding the Yolov3 python wrapper, library and configuration files in 
order to do detect the objects using the RealSense's RGB stream

[/Client/Detection/Detector.py]

- Adding a new class whose role is to detect objects using the RealSense 
RGB stream, YoloV3 neural network and calculate their real 3D position 
from the camera. To continue, the user may watch the 3D "World" 
environment, in addition to the objects detected on real time (as a red 
square)